pr review fixes, + add dummy data into children context
