+++ frontend/src/context/UserRoleContext.tsx
@@ -1,8 +1,7 @@
 'use client';
 
-import React, { createContext, ReactNode, useContext, useEffect, useMemo, useState } from 'react';
+import { createContext, ReactNode, useContext, useEffect, useMemo, useState } from 'react';
 import authService from '@/services/auth.service';
-import { User } from '@/types';
 import { isUserRole, UserRole } from '@/types/enums';
 
 const USER_ROLE_KEY = 'userRole';
