+++ frontend/src/services/calendar.service.ts
@@ -1,4 +1,3 @@
-import { AttendanceStatus, EventType } from '@/types/enums';
 import dayjs from 'dayjs';
 import { apiPaths, apiUrl } from '@/config/api';
 import {
@@ -7,6 +6,7 @@ import {
   parseJson,
   unwrapData,
 } from '@/services/http.service';
+import { AbsenceEntry, CreateAbsencePayload, CreateEventPayload, EventEntry, UpdateEventPayload } from '@/types';
 
 const USE_DUMMY_CALENDAR_DATA = true;
 
@@ -29,46 +29,136 @@ function withQuery(path: string, query?: Record<string, QueryValue>): string {
   return queryString ? `${path}?${queryString}` : path;
 }
 
-export type AbsenceEntry = {
-  id: number;
-  childId: number;
-  firstName?: string;
-  lastName?: string;
-  date: string;
-  status: AttendanceStatus;
-  note?: string | null;
-  userId: number;
-};
+class CalendarService {
+  async getAbsencesByChild({childId, from, to}: {childId: number; from: string; to: string}): Promise<AbsenceEntry[]> {
+    if (USE_DUMMY_CALENDAR_DATA) {
+      return buildDummyAbsences({ childId, from });
+    }
 
-export type CreateAbsencePayload = {
-  childId: number;
-  from: string;
-  to: string;
-  status: AttendanceStatus;
-  note?: string;
-};
+    const response = await fetchWithAuth(
+      apiUrl(withQuery(`/absences/child/${childId}`, { from, to })),
+    );
+    const payload = await parseJson(response);
 
-export type EventEntry = {
-  id: number;
-  name: string;
-  time: string;
-  description: string;
-  date: string;
-  type: EventType;
-};
+    if (!response.ok) {
+      throw new Error(extractErrorMessage(payload, `Failed to load absences (${response.status})`));
+    }
 
-export type CreateEventPayload = {
-  groupId: number;
-  from: string;
-  to: string;
-  timeFrom: string;
-  timeTo: string;
-  name: string;
-  description?: string;
-  type: EventType;
-};
+    return unwrapData<AbsenceEntry[]>(payload, []);
+  }
 
-export type UpdateEventPayload = CreateEventPayload;
+  async getAbsencesByGroup({groupId, from, to}: {groupId: number; from: string; to: string}): Promise<AbsenceEntry[]> {
+    if (USE_DUMMY_CALENDAR_DATA) {
+      return buildDummyGroupAbsences({ from });
+    }
+
+    const response = await fetchWithAuth(
+      apiUrl(withQuery(`/absences/group/${groupId}`, { from, to })),
+    );
+    const payload = await parseJson(response);
+
+    if (!response.ok) {
+      throw new Error(extractErrorMessage(payload, `Failed to load absences (${response.status})`));
+    }
+
+    return unwrapData<AbsenceEntry[]>(payload, []);
+  }
+
+  async createAbsence(body: CreateAbsencePayload): Promise<AbsenceEntry[] | null> {
+    const response = await fetchWithAuth(apiUrl('/absences'), {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify(body),
+    });
+    const payload = await parseJson(response);
+
+    if (!response.ok) {
+      throw new Error(extractErrorMessage(payload, `Failed to create absence (${response.status})`));
+    }
+
+    return unwrapData<AbsenceEntry[] | null>(payload, null);
+  }
+
+  async getEventsByGroup({ from, to}: { from: string; to: string}): Promise<EventEntry[]> {
+    if (USE_DUMMY_CALENDAR_DATA) {
+      return buildDummyEvents({ from });
+    }
+
+    const response = await fetchWithAuth(
+      apiUrl(apiPaths.events.listByRange({ from, to })),
+    );
+    const payload = await parseJson(response);
+
+    if (!response.ok) {
+      return [];
+    }
+
+    return unwrapData<EventEntry[]>(payload, []);
+  }
+
+  async getEventsByChild({childId, from, to}: {childId: number; from: string; to: string}): Promise<EventEntry[]> {
+    if (USE_DUMMY_CALENDAR_DATA) {
+      return buildDummyEvents({ from });
+    }
+
+    const response = await fetchWithAuth(
+      apiUrl(apiPaths.events.childByRange({ childId, from, to })),
+    );
+    const payload = await parseJson(response);
+
+    if (!response.ok) {
+      return [];
+    }
+
+    return unwrapData<EventEntry[]>(payload, []);
+  }
+
+  async createEvent(body: CreateEventPayload): Promise<EventEntry | null> {
+    const response = await fetchWithAuth(apiUrl(apiPaths.events.create), {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify(body),
+    });
+    const payload = await parseJson(response);
+
+    if (!response.ok) {
+      throw new Error(extractErrorMessage(payload, `Failed to create event (${response.status})`));
+    }
+
+    return unwrapData<EventEntry | null>(payload, null);
+  }
+
+  async updateEvent(eventId: number, body: UpdateEventPayload): Promise<EventEntry | null> {
+    const response = await fetchWithAuth(apiUrl(apiPaths.events.update(eventId)), {
+      method: 'PUT',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify(body),
+    });
+    const payload = await parseJson(response);
+
+    if (!response.ok) {
+      throw new Error(extractErrorMessage(payload, `Failed to update event (${response.status})`));
+    }
+
+    return unwrapData<EventEntry | null>(payload, null);
+  }
+
+  async deleteEvent(eventId: number): Promise<void> {
+    const response = await fetchWithAuth(apiUrl(apiPaths.events.remove(eventId)), {
+      method: 'DELETE',
+    });
+    const payload = await parseJson(response);
+
+    if (!response.ok) {
+      throw new Error(extractErrorMessage(payload, `Failed to delete event (${response.status})`));
+    }
+  }
+}
+
+const calendarService = new CalendarService();
+
+export default calendarService;
+
 
 function buildDummyAbsences({ childId, from }: { childId: number; from: string }): AbsenceEntry[] {
   const month = from.slice(0, 7);
@@ -214,133 +304,3 @@ function buildDummyEvents({ from }: { from: string }): EventEntry[] {
 
   return rows;
 }
-
-class CalendarService {
-  async getAbsencesByChild({childId, from, to}: {childId: number; from: string; to: string}): Promise<AbsenceEntry[]> {
-    if (USE_DUMMY_CALENDAR_DATA) {
-      return buildDummyAbsences({ childId, from });
-    }
-
-    const response = await fetchWithAuth(
-      apiUrl(withQuery(`/absences/child/${childId}`, { from, to })),
-    );
-    const payload = await parseJson(response);
-
-    if (!response.ok) {
-      throw new Error(extractErrorMessage(payload, `Failed to load absences (${response.status})`));
-    }
-
-    return unwrapData<AbsenceEntry[]>(payload, []);
-  }
-
-  async getAbsencesByGroup({groupId, from, to}: {groupId: number; from: string; to: string}): Promise<AbsenceEntry[]> {
-    if (USE_DUMMY_CALENDAR_DATA) {
-      return buildDummyGroupAbsences({ from });
-    }
-
-    const response = await fetchWithAuth(
-      apiUrl(withQuery(`/absences/group/${groupId}`, { from, to })),
-    );
-    const payload = await parseJson(response);
-
-    if (!response.ok) {
-      throw new Error(extractErrorMessage(payload, `Failed to load absences (${response.status})`));
-    }
-
-    return unwrapData<AbsenceEntry[]>(payload, []);
-  }
-
-  async createAbsence(body: CreateAbsencePayload): Promise<AbsenceEntry[] | null> {
-    const response = await fetchWithAuth(apiUrl('/absences'), {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify(body),
-    });
-    const payload = await parseJson(response);
-
-    if (!response.ok) {
-      throw new Error(extractErrorMessage(payload, `Failed to create absence (${response.status})`));
-    }
-
-    return unwrapData<AbsenceEntry[] | null>(payload, null);
-  }
-
-  async getEventsByGroup({ from, to}: { from: string; to: string}): Promise<EventEntry[]> {
-    if (USE_DUMMY_CALENDAR_DATA) {
-      return buildDummyEvents({ from });
-    }
-
-    const response = await fetchWithAuth(
-      apiUrl(apiPaths.events.listByRange({ from, to })),
-    );
-    const payload = await parseJson(response);
-
-    if (!response.ok) {
-      return [];
-    }
-
-    return unwrapData<EventEntry[]>(payload, []);
-  }
-
-  async getEventsByChild({childId, from, to}: {childId: number; from: string; to: string}): Promise<EventEntry[]> {
-    if (USE_DUMMY_CALENDAR_DATA) {
-      return buildDummyEvents({ from });
-    }
-
-    const response = await fetchWithAuth(
-      apiUrl(apiPaths.events.childByRange({ childId, from, to })),
-    );
-    const payload = await parseJson(response);
-
-    if (!response.ok) {
-      return [];
-    }
-
-    return unwrapData<EventEntry[]>(payload, []);
-  }
-
-  async createEvent(body: CreateEventPayload): Promise<EventEntry | null> {
-    const response = await fetchWithAuth(apiUrl(apiPaths.events.create), {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify(body),
-    });
-    const payload = await parseJson(response);
-
-    if (!response.ok) {
-      throw new Error(extractErrorMessage(payload, `Failed to create event (${response.status})`));
-    }
-
-    return unwrapData<EventEntry | null>(payload, null);
-  }
-
-  async updateEvent(eventId: number, body: UpdateEventPayload): Promise<EventEntry | null> {
-    const response = await fetchWithAuth(apiUrl(apiPaths.events.update(eventId)), {
-      method: 'PUT',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify(body),
-    });
-    const payload = await parseJson(response);
-
-    if (!response.ok) {
-      throw new Error(extractErrorMessage(payload, `Failed to update event (${response.status})`));
-    }
-
-    return unwrapData<EventEntry | null>(payload, null);
-  }
-
-  async deleteEvent(eventId: number): Promise<void> {
-    const response = await fetchWithAuth(apiUrl(apiPaths.events.remove(eventId)), {
-      method: 'DELETE',
-    });
-    const payload = await parseJson(response);
-
-    if (!response.ok) {
-      throw new Error(extractErrorMessage(payload, `Failed to delete event (${response.status})`));
-    }
-  }
-}
-
-const calendarService = new CalendarService();
-
-export default calendarService;
